refactor: :art: Transfered clerk auth check to server_util file
